Change keying function design
Go from providing both the input and output buffers to read/write to
providing only the input (ptr and len) and returning only the output
ptr (with the len provided as before).

That means the complexity is greatly reduced and more idiomatic. It also
enables zero-copy keying by allowing the function to return a pointer
within the input if wanted.

The error returning is not terribly impacted, with negative values being
error codes, in what will likely feel very familiar to C developers.

This changeset also introduces an alternative to the KEY_LENGTH (newly
renamed from key_length) global, in the form of a function named
key_length (in lowercase) which returns the correct value.

This helps greatly for some sources which cannot — or not properly —
produce exported globals.

All the types in the interface have also been tied down to precise types
instead of being left loose. This improves the implementation ease as
well as make the specification more robust and easier to understand.